usap-smc/pyproject.toml
[project]
name = "usap_smc"
version = "0.0.1"
description = "USAP Slice Manager and Classifier"
authors = [
    {name = "Murilo Silva",email = "murilosilva@itec.ufpa.br"},
    {name = "Victor Dias",email = "victor.leiter@ig.ufpa.br"},
    {name = "João Batista",email = "joao.bastos.junior@itec.ufpa.br"}
]
maintainers = [
    { name = "Murilo Silva", email = "murilosilva@itec.ufpa.br" },
    {name = "Victor Dias",email = "victor.leiter@ig.ufpa.br"},
    {name = "Matheus Costa",email = "matheus.gomes@itec.ufpa.br"},
    {name = "João Batista",email = "joao.bastos.junior@itec.ufpa.br"},
    {name = "Glauber Castro",email = "glaubersantiagodecastro@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = ["jsonify (>=0.5,<0.6)", "grpcio (>=1.69.0,<2.0.0)", "grpcio-tools (>=1.69.0,<2.0.0)", "pymongo (>=4.10.1,<5.0.0)", "pandas (>=2.2.3,<3.0.0)", "keras-tuner (>=1.4.7,<2.0.0)", "scikit-learn (>=1.6.1,<2.0.0)", "tensorflow-cpu (>=2.18.0,<3.0.0)", "loguru (>=0.7.3,<0.8.0)"]
license = { text = "Apache-2.0" }

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
requires-poetry = ">=2.0"
classifiers = [
    "Topic :: Software Development :: Open RAN :: Network Slicing",
    "Topic :: Software Development :: Open RAN :: Monitoring",
    "Topic :: Software Development :: Open RAN :: Machine Learning",
    "Topic :: Software Development :: Open RAN :: Artificial Intelligence (AI)"
]

[project.scripts]
usap_smc = "usap_smc.main:run"